Added branch_1 Переименованы классы и файлы. Ввод смс кода вынесен в отдельный экран
