+++ lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:firtsflutterapp/SignInPage.dart';
+import 'file:///D:/Programming/Flutter/firts_flutter_app/lib/auth/PhoneInput.dart';
 import 'package:flutter/material.dart';
 
 void main() {
@@ -18,7 +18,7 @@ class MyApp extends StatelessWidget {
 
         primarySwatch: Colors.blue,
       ),
-      home: SignInPage(),
+      home: PhoneInput(),
     );
   }
 }
